+++ wtfix/core/tests/test_encoders.py
@@ -4,14 +4,14 @@ import pytest
 
 from wtfix.core import encoders, decoders
 from wtfix.core.exceptions import DuplicateTags
-from wtfix.message.fieldset import OrderedDictFieldSet, ListFieldSet
+from wtfix.message.collections import FieldDict, FieldList
 
 
 class TestJSONMessageEncoder:
-    def test_to_json_encodes_nested_orderedfieldset_as_expected(
+    def test_to_json_encodes_nested_fielddict_as_expected(
         self, nested_parties_group, encoded_dict_sample
     ):
-        fs = OrderedDictFieldSet(
+        fs = FieldDict(
             (1, "a"),
             (2, "b"),
             nested_parties_group.identifier,
@@ -22,11 +22,11 @@ class TestJSONMessageEncoder:
 
         assert encoders.to_json(fs) == encoded_dict_sample
 
-    def test_to_json_orderedfieldset_raises_exception_on_duplicate_tags_without_template_defined(
+    def test_to_json_fielddict_raises_exception_on_duplicate_tags_without_template_defined(
         self, nested_parties_group
     ):
         with pytest.raises(DuplicateTags):
-            fs = OrderedDictFieldSet(
+            fs = FieldDict(
                 (1, "a"),
                 (2, "b"),
                 nested_parties_group.identifier,
@@ -36,10 +36,10 @@ class TestJSONMessageEncoder:
 
             encoders.to_json(fs)
 
-    def test_to_json_nested_listfieldset_encodes_as_expected(
+    def test_to_json_nested_fieldlist_encodes_as_expected(
         self, nested_parties_group, encoded_list_sample
     ):
-        fs = ListFieldSet(
+        fs = FieldList(
             (1, "a"),
             (2, "b"),
             nested_parties_group.identifier,
@@ -52,7 +52,7 @@ class TestJSONMessageEncoder:
     def test_to_json_encodes_bytestrings(
         self, nested_parties_group, encoded_list_sample
     ):
-        fs = ListFieldSet(
+        fs = FieldList(
             (1, "a"),
             (2, b"b"),
             (3, "c"),
@@ -61,7 +61,7 @@ class TestJSONMessageEncoder:
         assert encoders.to_json(fs) == json.dumps([[1, "a"], [2, "b"], [3, "c"]])
 
 
-def test_serialization_is_idempotent(fieldset_class, nested_parties_group):
+def test_serialization_is_idempotent(fieldmap_class, nested_parties_group):
     kwargs = {}
     fields = [
         (1, "a"),
@@ -71,10 +71,10 @@ def test_serialization_is_idempotent(fieldset_class, nested_parties_group):
         (3, "c"),
     ]
 
-    if fieldset_class.__name__ == OrderedDictFieldSet.__name__:
+    if fieldmap_class.__name__ == FieldDict.__name__:
         kwargs["group_templates"] = {539: [524, 525, 538, 804], 804: [545, 805]}
 
-    fs = fieldset_class(*fields, **kwargs)
+    fs = fieldmap_class(*fields, **kwargs)
 
     encoded = encoders.to_json(fs)
     decoded = decoders.from_json(encoded)